refactor: (31.03.24) Moved `OpenScheduleMessageDialogButton` from `Message` to `MessageForm` (#221)
