+++ client/src/Message/components/index.ts
@@ -12,7 +12,6 @@ export * from "./PinMessageSnackbarManager";
 export * from "./PinnedMessage";
 export * from "./UnpinMessageSnackbarManager";
 export * from "./ScheduleMessageDialog";
-export * from "./OpenScheduleMessageDialogButton";
 export * from "./ScheduledMessagesAppBar";
 export * from "./ScheduledMessagesList";
 export * from "./ScheduledMessageMenu";
