+++ client/src/MessageForm/components/PlainTextMessageForm.tsx
@@ -6,9 +6,9 @@ import {EmojiData} from "emoji-mart";
 import "emoji-mart/css/emoji-mart.css";
 import {AttachFilesButton} from "./AttachFilesButton";
 import {SendMessageButton} from "./SendMessageButton";
+import {OpenScheduleMessageDialogButton} from "./OpenScheduleMessageDialogButton";
 import {RecordVoiceMessageButton} from "./RecordVoiceMessageButton";
 import {MessageFormData} from "../types";
-import {OpenScheduleMessageDialogButton} from "../../Message";
 import {EmojiAndStickerPicker, EmojiPickerContainer} from "../../EmojiPicker";
 import {MarkdownPreviewDialog} from "../../Markdown";
 import {useLocalization, useStore} from "../../store";
